+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.1.1
+ - Add "Unselect all" button
+ - Fix bug where select all button was imply triggering a toggle event on every row, not turning them on
+
 ## v1.1.0
  - Add support to filter by mod tier
  - Add "select all filtered" option